--- src/main/java/wanted/media/content/domain/CountValueType.java
@@ -1,5 +1,0 @@
-package wanted.media.content.domain;
-
-public enum CountValueType {
-	COUNT, LIKE_COUNT, VIEW_COUNT, SHARE_COUNT;
-}
