--- src/main/java/wanted/media/content/domain/StatDateType.java
@@ -1,5 +1,0 @@
-package wanted.media.content.domain;
-
-public enum StatDateType {
-	DATE, HOUR;
-}
