+++ src/main/java/wanted/media/post/domain/CountValueType.java
@@ -1,0 +1,5 @@
+package wanted.media.post.domain;
+
+public enum CountValueType {
+    COUNT, LIKE_COUNT, VIEW_COUNT, SHARE_COUNT;
+}
