+++ src/main/java/wanted/media/post/domain/StatDateType.java
@@ -1,0 +1,5 @@
+package wanted.media.post.domain;
+
+public enum StatDateType {
+    DATE, HOUR;
+}
